refactor(runtime): split backend source layers
