+++ components/aihc-amd64/aihc-amd64.cabal
@@ -46,6 +46,7 @@ test-suite spec
     aihc-amd64,
     aihc-fc,
     aihc-grin,
+    aihc-native,
     aihc-parser,
     aihc-resolve,
     aihc-tc,
