+++ components/aihc-amd64/src/Aihc/Amd64/Codegen.hs
@@ -1887,7 +1887,7 @@ renderObservedMetadata env program resultReps = do
   constructors <- mapM renderConstructorDescriptor constructorEntries
   functions <- mapM renderFunctionDescriptor functionEntries
   pure . T.unlines $
-    [ "#include \"aihc_runtime.h\"",
+    [ "#include \"aihc_snapshot.h\"",
       "#include <stddef.h>",
       ""
     ]
