+++ components/aihc-arm64/aihc-arm64.cabal
@@ -46,6 +46,7 @@ test-suite spec
     aihc-arm64,
     aihc-fc,
     aihc-grin,
+    aihc-native,
     aihc-parser,
     aihc-resolve,
     aihc-tc,
