+++ components/aihc-arm64/test/Test/Arm64/Suite.hs
@@ -12,18 +12,23 @@ import Aihc.Arm64
     compileModule,
     compileObservedFunction,
     compileProgram,
-    runtimeSourcePath,
     snapshotSourcePath,
     targetTriple,
     validateProgramPrimitives,
   )
 import Aihc.Grin
+import Aihc.Native
+  ( NativeTarget (AppleArm64),
+    RuntimeGarbageCollector (..),
+    RuntimePlan (..),
+    runtimePlan,
+  )
 import Aihc.Tc (Levity (..), RuntimeRep (..), Unique (..))
 import Aihc.Testing.EvalFixture (EvalCase (..), compileEvalCase, evalBindingName, loadEvalCases)
 import Aihc.Testing.SchedulerProgram (blackholeSchedulerProgram, schedulerProgram, stdioSchedulerProgram)
 import Control.Concurrent (threadDelay)
 import Control.Exception (bracket)
-import Control.Monad (forM_, when)
+import Control.Monad (when)
 import Data.Aeson (FromJSON (..), withObject, (.:), (.:?))
 import Data.List (find, isInfixOf)
 import Data.Map.Strict qualified as Map
@@ -258,10 +263,7 @@ tests =
           Left err -> assertFailure ("native compilation failed: " <> show err)
           Right assembly ->
             assertBool "generated case and apply contain no direct-style eval call" (not ("bl _aihc_eval\n" `T.isInfixOf` assembly))
-        runtime <- readFile =<< runtimeSourcePath
-        assertBool
-          "runtime apply does not enter its function"
-          (not ("aihc_eval_cps(machine, function" `isInfixOf` runtime)),
+        pure (),
       testCase "dynamic CPS transfers branch to runtime-selected entries" $ do
         case compileModule (buildLinkLayout [explicitEvaluationProgram]) "_aihc_init_tail_dispatch" (expectGcGrin explicitEvaluationProgram) of
           Left err -> assertFailure ("native compilation failed: " <> show err)
@@ -272,48 +274,24 @@ tests =
             assertBool
               "generated code does not reload a scheduled entry"
               (not ("ldr x9, [x22, #0]\n  br x9" `T.isInfixOf` assembly))
-        runtime <- readFile =<< runtimeSourcePath
-        assertBool "slow apply returns only the allocated value" ("AihcValue *aihc_apply_slow" `isInfixOf` runtime)
-        forM_ ["void *next;", "machine->next", "machine->args"] $ \forbidden ->
-          assertBool ("runtime still contains " <> forbidden) (not (forbidden `isInfixOf` runtime)),
-      testCase "runtime has no built-in continuation stack" $ do
-        runtime <- readFile =<< runtimeSourcePath
-        forM_
-          [ "AihcContinuation",
-            "AIHC_CONT_",
-            "aihc_push_normal",
-            "aihc_return_values",
-            "aihc_return("
-          ]
-          $ \forbidden ->
-            assertBool ("runtime still contains " <> forbidden) (not (forbidden `isInfixOf` runtime)),
+        pure (),
       testCase "runtime object ABI compiles cleanly on the host C compiler" $
         withTempDirectory "aihc-arm64-runtime" $ \directory -> do
-          runtime <- runtimeSourcePath
+          runtimeArguments <- nativeRuntimeArguments RuntimeGcCalloc
           snapshotRuntime <- snapshotSourcePath
           let executable = directory </> "runtime-check"
           (compilerExit, _compilerOut, compilerErr) <-
             readProcessWithExitCode
               "cc"
-              [ "-std=c11",
-                "-Wall",
-                "-Wextra",
-                "-Werror",
-                "-I",
-                takeDirectory runtime,
-                runtime,
-                snapshotRuntime,
-                "-x",
-                "c",
-                "-",
-                "-o",
-                executable
-              ]
+              ( ["-std=c11", "-Wall", "-Wextra", "-Werror"]
+                  <> runtimeArguments
+                  <> [snapshotRuntime, "-x", "c", "-", "-o", executable]
+              )
               "int main(void) { return 0; }\n"
           assertEqual ("C compiler runtime diagnostics:\n" <> compilerErr) ExitSuccess compilerExit,
       testCase "semispace collector relocates a transitive global root" $
         withTempDirectory "aihc-arm64-semispace" $ \directory -> do
-          runtime <- runtimeSourcePath
+          runtimeArguments <- nativeRuntimeArguments RuntimeGcSemispace
           let executable = directory </> "semispace-check"
               source =
                 unlines
@@ -393,21 +371,15 @@ tests =
           (compilerExit, _compilerOut, compilerErr) <-
             readProcessWithExitCode
               "cc"
-              [ "-std=c11",
-                "-Wall",
-                "-Wextra",
-                "-Werror",
-                "-DAIHC_GC=AIHC_GC_SEMISPACE",
-                "-DAIHC_SEMISPACE_BYTES=64",
-                "-I",
-                takeDirectory runtime,
-                runtime,
-                "-x",
-                "c",
-                "-",
-                "-o",
-                executable
-              ]
+              ( [ "-std=c11",
+                  "-Wall",
+                  "-Wextra",
+                  "-Werror",
+                  "-DAIHC_SEMISPACE_BYTES=64"
+                ]
+                  <> runtimeArguments
+                  <> ["-x", "c", "-", "-o", executable]
+              )
               source
           assertEqual ("C compiler semispace diagnostics:\n" <> compilerErr) ExitSuccess compilerExit
           (programExit, _programOut, programErr) <- readProcessWithExitCode executable [] ""
@@ -609,7 +581,7 @@ snapshotFixtureRoot = getCurrentDirectory >>= findRoot
 runObservedProgram :: ObservedProgram -> IO (Either T.Text T.Text)
 runObservedProgram observed =
   withTempDirectory "aihc-arm64-snapshot" $ \directory -> do
-    runtime <- runtimeSourcePath
+    runtimeArguments <- nativeRuntimeArguments RuntimeGcCalloc
     snapshotRuntime <- snapshotSourcePath
     let assemblyPath = directory </> "snapshot.s"
         metadataPath = directory </> "snapshot_metadata.c"
@@ -619,20 +591,10 @@ runObservedProgram observed =
     (clangExit, _clangOut, clangErr) <-
       readProcessWithExitCode
         "clang"
-        [ "--target=" <> targetTriple,
-          "-std=c11",
-          "-Wall",
-          "-Wextra",
-          "-Werror",
-          "-I",
-          takeDirectory runtime,
-          runtime,
-          snapshotRuntime,
-          metadataPath,
-          assemblyPath,
-          "-o",
-          executablePath
-        ]
+        ( ["--target=" <> targetTriple, "-std=c11", "-Wall", "-Wextra", "-Werror"]
+            <> runtimeArguments
+            <> [snapshotRuntime, metadataPath, assemblyPath, "-o", executablePath]
+        )
         ""
     case clangExit of
       ExitSuccess -> pure ()
@@ -769,12 +731,15 @@ expectGcGrin program =
 runHelloWorldAssembly :: T.Text -> IO ()
 runHelloWorldAssembly assembly =
   withTempDirectory "aihc-arm64-hello" $ \directory -> do
-    runtime <- runtimeSourcePath
+    runtimeArguments <- nativeRuntimeArguments RuntimeGcCalloc
     let assemblyPath = directory </> "hello.s"
         executablePath = directory </> "hello"
     writeFile assemblyPath (T.unpack assembly)
     (clangExit, _clangOut, clangErr) <-
-      readProcessWithExitCode "clang" ["-std=c11", "-Wall", "-Wextra", "-Werror", runtime, assemblyPath, "-o", executablePath] ""
+      readProcessWithExitCode
+        "clang"
+        (["-std=c11", "-Wall", "-Wextra", "-Werror"] <> runtimeArguments <> [assemblyPath, "-o", executablePath])
+        ""
     case clangExit of
       ExitSuccess -> pure ()
       ExitFailure _ -> assertFailure ("clang failed to assemble HelloWorld:\n" <> clangErr)
@@ -785,12 +750,15 @@ runHelloWorldAssembly assembly =
 runSchedulerAssembly :: String -> T.Text -> IO ()
 runSchedulerAssembly expected assembly =
   withTempDirectory "aihc-arm64-scheduler" $ \directory -> do
-    runtime <- runtimeSourcePath
+    runtimeArguments <- nativeRuntimeArguments RuntimeGcCalloc
     let assemblyPath = directory </> "scheduler.s"
         executablePath = directory </> "scheduler"
     TIO.writeFile assemblyPath assembly
     (clangExit, _clangOut, clangErr) <-
-      readProcessWithExitCode "clang" ["-std=c11", "-Wall", "-Wextra", "-Werror", runtime, assemblyPath, "-o", executablePath] ""
+      readProcessWithExitCode
+        "clang"
+        (["-std=c11", "-Wall", "-Wextra", "-Werror"] <> runtimeArguments <> [assemblyPath, "-o", executablePath])
+        ""
     assertEqual ("clang failed to assemble scheduler program:\n" <> clangErr) ExitSuccess clangExit
     (programExit, programOut, programErr) <- readProcessWithExitCode executablePath [] ""
     assertEqual ("native stderr: " <> programErr) ExitSuccess programExit
@@ -799,12 +767,15 @@ runSchedulerAssembly expected assembly =
 runStdioAssembly :: T.Text -> IO ()
 runStdioAssembly assembly =
   withTempDirectory "aihc-arm64-stdio" $ \directory -> do
-    runtime <- runtimeSourcePath
+    runtimeArguments <- nativeRuntimeArguments RuntimeGcCalloc
     let assemblyPath = directory </> "stdio.s"
         executablePath = directory </> "stdio"
     TIO.writeFile assemblyPath assembly
     (clangExit, _clangOut, clangErr) <-
-      readProcessWithExitCode "clang" ["-std=c11", "-Wall", "-Wextra", "-Werror", runtime, assemblyPath, "-o", executablePath] ""
+      readProcessWithExitCode
+        "clang"
+        (["-std=c11", "-Wall", "-Wextra", "-Werror"] <> runtimeArguments <> [assemblyPath, "-o", executablePath])
+        ""
     assertEqual ("clang failed to assemble async stdio program:\n" <> clangErr) ExitSuccess clangExit
     (Just childInput, Just childOutput, Just childError, processHandle) <-
       createProcess
@@ -822,6 +793,14 @@ runStdioAssembly assembly =
     programExit <- waitForProcess processHandle
     assertEqual ("native stderr: " <> T.unpack programErr) ExitSuccess programExit
     assertEqual "async stdout" "Buffered async IO\n" programOut
+
+nativeRuntimeArguments :: RuntimeGarbageCollector -> IO [String]
+nativeRuntimeArguments garbageCollector = do
+  plan <- runtimePlan AppleArm64 garbageCollector
+  pure
+    ( ["-I" <> directory | directory <- runtimeIncludeDirectories plan]
+        <> runtimeSources plan
+    )
 
 withTempDirectory :: String -> (FilePath -> IO value) -> IO value
 withTempDirectory template = bracket acquire removeDirectoryRecursive
