+++ components/aihc-native/aihc-native.cabal
@@ -9,9 +9,17 @@ maintainer: lemmih@gmail.com
 category: Development
 synopsis: Shared backend compilation support for AIHC
 data-files:
+  runtime/aihc_gc_calloc.c
+  runtime/aihc_gc_semispace.c
+  runtime/aihc_host_posix.c
+  runtime/aihc_host_wasip3.c
   runtime/aihc_runtime.c
   runtime/aihc_runtime.h
+  runtime/aihc_runtime_internal.h
+  runtime/aihc_runtime_trampoline.c
+  runtime/aihc_runtime_trampoline.h
   runtime/aihc_snapshot.c
+  runtime/aihc_snapshot.h
 
 library
   exposed-modules:
@@ -28,6 +36,7 @@ library
     base >=4.16 && <5,
     bytestring,
     containers,
+    filepath,
     text,
 
   ghc-options: -Wall
@@ -41,6 +50,7 @@ test-suite spec
     Test.Native.BlockLayout
     Test.Native.Primitive
     Test.Native.RegisterAllocate
+    Test.Native.RuntimePlan
 
   build-depends:
     aihc-grin,
@@ -48,6 +58,7 @@ test-suite spec
     aihc-tc,
     base >=4.16 && <5,
     containers,
+    filepath,
     tasty,
     tasty-hunit,
     tasty-quickcheck,
