+++ components/aihc-native/runtime/aihc_runtime.c
@@ -1,120 +1,17 @@
 #include "aihc_runtime.h"
+#include "aihc_runtime_internal.h"
 
 #include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
-#ifndef AIHC_WASIP3
-#include <errno.h>
-#include <fcntl.h>
-#include <poll.h>
-#include <stdio.h>
-#include <unistd.h>
-#endif
-
-typedef struct AihcBlackholeWaiter AihcBlackholeWaiter;
-typedef struct AihcMVarWaiter AihcMVarWaiter;
-
-typedef enum {
-  AIHC_IO_READ,
-  AIHC_IO_WRITE,
-} AihcIoKind;
-
-typedef enum {
-  AIHC_IO_READABLE = 1U << 0,
-  AIHC_IO_WRITABLE = 1U << 1,
-} AihcIoCapability;
-
-typedef enum {
-  AIHC_IO_SUBMITTED,
-  AIHC_IO_PENDING,
-  AIHC_IO_COMPLETED,
-  AIHC_IO_CONSUMED,
-} AihcIoState;
-
-struct AihcThread {
-  AihcSlot header;
-  AihcResumeKind resume_kind;
-  AihcValue *resume_function;
-  AihcValue *resume_continuation;
-  AihcSlot resume_value;
-  uint64_t resume_count;
-  AihcThread *next;
-};
-
-struct AihcBlackholeWaiter {
-  AihcThread *thread;
-  AihcValue *continuation;
-  AihcBlackholeWaiter *next;
-};
-
-struct AihcBlackhole {
-  AihcValue *object;
-  AihcThread *owner;
-  AihcBlackholeWaiter *waiters_head;
-  AihcBlackholeWaiter *waiters_tail;
-  AihcBlackhole *next;
-};
-
-struct AihcMVarWaiter {
-  AihcThread *thread;
-  AihcValue *continuation;
-  AihcSlot value;
-  AihcMVarWaiter *next;
-};
-
-struct AihcMVar {
-  uint8_t full;
-  AihcSlot value;
-  AihcMVarWaiter *readers_head;
-  AihcMVarWaiter *readers_tail;
-  AihcMVarWaiter *takers_head;
-  AihcMVarWaiter *takers_tail;
-  AihcMVarWaiter *putters_head;
-  AihcMVarWaiter *putters_tail;
-  AihcMVar *next;
-};
-
-struct AihcIoHandle {
-  uintptr_t backend_token;
-  uint32_t capabilities;
-};
-
-typedef struct {
-  size_t size;
-  uint8_t *contents;
-  uint8_t pinned;
-  size_t alignment;
-} AihcByteArray;
-
-struct AihcIoRequest {
-  AihcIoKind kind;
-  AihcIoState state;
-  AihcIoHandle *handle;
-  uint8_t *buffer;
-  size_t offset;
-  size_t length;
-  AihcThread *thread;
-  AihcValue *continuation;
-  int32_t result;
-  AihcIoRequest *next;
-};
-
-struct AihcIoBackend {
-  int (*prepare)(AihcIoRequest *request);
-  int (*try_request)(AihcIoRequest *request, int32_t *result);
-  void (*poll)(AihcMachine *machine, int may_block);
-};
-
-static AihcIoHandle aihc_standard_input = {(uintptr_t)0, AIHC_IO_READABLE};
-static AihcIoHandle aihc_standard_output = {(uintptr_t)1, AIHC_IO_WRITABLE};
 
 #if UINTPTR_MAX == UINT64_MAX
 _Static_assert(offsetof(AihcMachine, globals) == 0, "machine globals ABI");
 _Static_assert(offsetof(AihcMachine, heap_next) == 24, "machine heap-next ABI");
 _Static_assert(offsetof(AihcMachine, heap_limit) == 32,
                "machine heap-limit ABI");
-_Static_assert(offsetof(AihcInfo, enter_entry) == 48,
-               "info-table enter-entry ABI");
+_Static_assert(offsetof(AihcInfo, backend_entry) == 48,
+               "info-table backend-entry ABI");
 _Static_assert(offsetof(AihcResume, kind) == 0, "resume kind ABI");
 _Static_assert(offsetof(AihcResume, function) == 8, "resume function ABI");
 _Static_assert(offsetof(AihcResume, continuation) == 16,
@@ -123,31 +20,22 @@ _Static_assert(offsetof(AihcResume, value) == 24, "resume value ABI");
 _Static_assert(offsetof(AihcResume, count) == 32, "resume count ABI");
 #endif
 
-static _Noreturn void aihc_fail(const char *message) {
-#ifdef AIHC_WASIP3
-  (void)message;
-  __builtin_trap();
-#else
-  fprintf(stderr, "aihc runtime: %s\n", message);
-  abort();
-#endif
-}
+_Noreturn void aihc_fail(const char *message) { aihc_host_fail(message); }
 
 static const AihcResume *aihc_schedule(AihcMachine *machine);
-static const AihcIoBackend *aihc_default_io_backend(void);
 
 void aihc_unsupported_primitive(void) {
   aihc_fail("primitive is not implemented by the native runtime");
 }
 
-static void aihc_record_allocation(AihcMachine *machine) {
+void aihc_record_allocation(AihcMachine *machine) {
   if (machine->allocation_count == UINT64_MAX) {
     aihc_fail("allocation counter overflow");
   }
   ++machine->allocation_count;
 }
 
-static void *aihc_allocate_zeroed(size_t bytes) {
+void *aihc_allocate_zeroed(size_t bytes) {
   void *pointer = calloc(1, bytes);
   if (pointer == NULL) {
     aihc_fail("out of memory");
@@ -155,14 +43,14 @@ static void *aihc_allocate_zeroed(size_t bytes) {
   return pointer;
 }
 
-static void *aihc_allocate_auxiliary(AihcMachine *machine, size_t bytes) {
+void *aihc_allocate_auxiliary(AihcMachine *machine, size_t bytes) {
   void *pointer = aihc_allocate_zeroed(bytes);
   aihc_record_allocation(machine);
   return pointer;
 }
 
-static AihcSlot *aihc_reserve_slots(AihcMachine *machine, AihcSlot **slots,
-                                    uint64_t *capacity, uint64_t count) {
+AihcSlot *aihc_reserve_slots(AihcMachine *machine, AihcSlot **slots,
+                             uint64_t *capacity, uint64_t count) {
   uint64_t required = count == 0 ? 1 : count;
   if (required <= *capacity) {
     return *slots;
@@ -199,7 +87,7 @@ static AihcSlot aihc_make_header(uint64_t tag, const AihcInfo *info) {
   }
 }
 
-static uint64_t aihc_object_words(uint64_t tag, const AihcInfo *info) {
+uint64_t aihc_object_words(uint64_t tag, const AihcInfo *info) {
   uint64_t field_words = info->field_count;
   if (field_words == 0 && (tag == AIHC_TAG_THUNK || tag == AIHC_TAG_BLACKHOLE ||
                            tag == AIHC_TAG_INDIRECTION)) {
@@ -208,187 +96,88 @@ static uint64_t aihc_object_words(uint64_t tag, const AihcInfo *info) {
   return 1 + field_words;
 }
 
-#if AIHC_GC == AIHC_GC_SEMISPACE
-static int aihc_in_space(const uint8_t *start, uint64_t bytes,
-                         const AihcValue *value) {
-  uintptr_t address = (uintptr_t)value;
-  uintptr_t first = (uintptr_t)start;
-  return address >= first && address - first < bytes;
+static void aihc_visit_value(AihcValue **value, AihcRootVisitor visitor,
+                             void *context) {
+  *value =
+      (AihcValue *)(uintptr_t)visitor((AihcSlot)(uintptr_t)*value, context);
 }
 
-static AihcValue *aihc_forward(AihcMachine *machine, uint8_t *from_start,
-                               AihcValue *value) {
-  if (value == NULL ||
-      !aihc_in_space(from_start, machine->semispace_bytes, value)) {
-    return value;
+static void aihc_visit_thread(AihcThread *thread, AihcRootVisitor visitor,
+                              void *context) {
+  if (thread == NULL) {
+    return;
   }
-  if (aihc_value_tag(value) == AIHC_TAG_FORWARDING) {
-    return (AihcValue *)(value->header & ~AIHC_TAG_MASK);
-  }
-
-  uint64_t tag = aihc_value_tag(value);
-  const AihcInfo *info = aihc_value_info_table(value);
-  uint64_t words = aihc_object_words(tag, info);
-  size_t bytes = sizeof(AihcSlot) * words;
-  if (machine->heap_next + bytes > machine->heap_limit) {
-    aihc_fail("live data exceeds semispace");
-  }
-  AihcValue *copy = (AihcValue *)machine->heap_next;
-  memcpy(copy, value, bytes);
-  machine->heap_next += bytes;
-  value->header = (uintptr_t)copy | AIHC_TAG_FORWARDING;
-  return copy;
-}
-
-static void aihc_forward_slot(AihcMachine *machine, uint8_t *from_start,
-                              AihcSlot *slot) {
-  *slot = (AihcSlot)aihc_forward(machine, from_start, (AihcValue *)*slot);
-}
-
-static void aihc_forward_value(AihcMachine *machine, uint8_t *from_start,
-                               AihcValue **value) {
-  *value = aihc_forward(machine, from_start, *value);
-}
-
-static void aihc_forward_thread(AihcMachine *machine, uint8_t *from_start,
-                                AihcThread *thread) {
-  if (thread != NULL) {
-    aihc_forward_value(machine, from_start, &thread->resume_function);
-    aihc_forward_value(machine, from_start, &thread->resume_continuation);
-    if (thread->resume_kind == AIHC_RESUME_CONTINUE &&
-        thread->resume_count == 1) {
-      aihc_forward_slot(machine, from_start, &thread->resume_value);
-    }
+  aihc_visit_value(&thread->resume_function, visitor, context);
+  aihc_visit_value(&thread->resume_continuation, visitor, context);
+  if (thread->resume_kind == AIHC_RESUME_CONTINUE &&
+      thread->resume_count == 1) {
+    thread->resume_value = visitor(thread->resume_value, context);
   }
 }
 
-static void aihc_collect(AihcMachine *machine, uint64_t required_words,
-                         uint64_t root_count, AihcSlot *roots) {
-  uint8_t *from_start = machine->heap_start;
-  uint8_t *to_start = machine->other_space;
-  machine->heap_start = to_start;
-  machine->heap_next = to_start;
-  machine->heap_limit = to_start + machine->semispace_bytes;
-
+void aihc_visit_roots(AihcMachine *machine, uint64_t root_count,
+                      AihcSlot *roots, AihcRootVisitor visitor, void *context) {
   for (uint64_t index = 0; index < machine->global_count; ++index) {
-    aihc_forward_slot(machine, from_start, &machine->globals[index]);
+    machine->globals[index] = visitor(machine->globals[index], context);
   }
   for (uint64_t index = 0; index < root_count; ++index) {
-    aihc_forward_slot(machine, from_start, &roots[index]);
+    roots[index] = visitor(roots[index], context);
   }
-  aihc_forward_value(machine, from_start, &machine->thread_done_continuation);
-  aihc_forward_value(machine, from_start, &machine->selected_resume.function);
-  aihc_forward_value(machine, from_start,
-                     &machine->selected_resume.continuation);
+  aihc_visit_value(&machine->thread_done_continuation, visitor, context);
+  aihc_visit_value(&machine->selected_resume.function, visitor, context);
+  aihc_visit_value(&machine->selected_resume.continuation, visitor, context);
   if (machine->selected_resume.kind == AIHC_RESUME_CONTINUE &&
       machine->selected_resume.count == 1) {
-    aihc_forward_slot(machine, from_start, &machine->selected_resume.value);
+    machine->selected_resume.value =
+        visitor(machine->selected_resume.value, context);
   }
-  aihc_forward_thread(machine, from_start, machine->current_thread);
+  aihc_visit_thread(machine->current_thread, visitor, context);
   for (AihcThread *thread = machine->run_queue_head; thread != NULL;
        thread = thread->next) {
-    aihc_forward_thread(machine, from_start, thread);
+    aihc_visit_thread(thread, visitor, context);
   }
   for (AihcBlackhole *blackhole = machine->blackholes; blackhole != NULL;
        blackhole = blackhole->next) {
-    aihc_forward_value(machine, from_start, &blackhole->object);
+    aihc_visit_value(&blackhole->object, visitor, context);
     for (AihcBlackholeWaiter *waiter = blackhole->waiters_head; waiter != NULL;
          waiter = waiter->next) {
-      aihc_forward_value(machine, from_start, &waiter->continuation);
-      aihc_forward_thread(machine, from_start, waiter->thread);
+      aihc_visit_value(&waiter->continuation, visitor, context);
+      aihc_visit_thread(waiter->thread, visitor, context);
     }
   }
   for (AihcMVar *mvar = machine->mvars; mvar != NULL; mvar = mvar->next) {
     if (mvar->full) {
-      aihc_forward_slot(machine, from_start, &mvar->value);
+      mvar->value = visitor(mvar->value, context);
     }
-    for (AihcMVarWaiter *waiter = mvar->readers_head; waiter != NULL;
-         waiter = waiter->next) {
-      aihc_forward_value(machine, from_start, &waiter->continuation);
-      aihc_forward_thread(machine, from_start, waiter->thread);
-    }
-    for (AihcMVarWaiter *waiter = mvar->takers_head; waiter != NULL;
-         waiter = waiter->next) {
-      aihc_forward_value(machine, from_start, &waiter->continuation);
-      aihc_forward_thread(machine, from_start, waiter->thread);
-    }
-    for (AihcMVarWaiter *waiter = mvar->putters_head; waiter != NULL;
-         waiter = waiter->next) {
-      aihc_forward_slot(machine, from_start, &waiter->value);
-      aihc_forward_value(machine, from_start, &waiter->continuation);
-      aihc_forward_thread(machine, from_start, waiter->thread);
+    AihcMVarWaiter *waiter_lists[] = {mvar->readers_head, mvar->takers_head,
+                                      mvar->putters_head};
+    for (size_t list = 0; list < 3; ++list) {
+      for (AihcMVarWaiter *waiter = waiter_lists[list]; waiter != NULL;
+           waiter = waiter->next) {
+        if (list == 2) {
+          waiter->value = visitor(waiter->value, context);
+        }
+        aihc_visit_value(&waiter->continuation, visitor, context);
+        aihc_visit_thread(waiter->thread, visitor, context);
+      }
     }
   }
   for (AihcIoRequest *request = machine->io_requests_head; request != NULL;
        request = request->next) {
-    aihc_forward_value(machine, from_start, &request->continuation);
-    aihc_forward_thread(machine, from_start, request->thread);
-  }
-
-  uint8_t *scan = to_start;
-  while (scan < machine->heap_next) {
-    AihcValue *object = (AihcValue *)scan;
-    uint64_t tag = aihc_value_tag(object);
-    const AihcInfo *info = aihc_value_info_table(object);
-    uint64_t count = info->field_count;
-    if (tag == AIHC_TAG_INDIRECTION) {
-      aihc_forward_slot(machine, from_start, &object->fields[0]);
-    } else if (tag == AIHC_TAG_NODE || tag == AIHC_TAG_CLOSURE ||
-               tag == AIHC_TAG_THUNK || tag == AIHC_TAG_PARTIAL_CONSTRUCTOR ||
-               tag == AIHC_TAG_BLACKHOLE) {
-      for (uint64_t index = 0; index < count; ++index) {
-        if (info->field_is_pointer != NULL && info->field_is_pointer[index]) {
-          aihc_forward_slot(machine, from_start, &object->fields[index]);
-        }
-      }
-    } else {
-      aihc_fail("collector encountered an invalid object tag");
-    }
-    scan += sizeof(AihcSlot) * aihc_object_words(tag, info);
-  }
-
-  machine->other_space = from_start;
-  size_t required_bytes = sizeof(AihcSlot) * required_words;
-  if (machine->heap_next + required_bytes > machine->heap_limit) {
-    aihc_fail("insufficient heap after collection");
+    aihc_visit_value(&request->continuation, visitor, context);
+    aihc_visit_thread(request->thread, visitor, context);
   }
 }
-#endif
 
 void aihc_ensure_heap(AihcMachine *machine, uint64_t words, uint64_t root_count,
                       AihcSlot *roots) {
-#if AIHC_GC == AIHC_GC_CALLOC
-  (void)machine;
-  (void)words;
-  (void)root_count;
-  (void)roots;
-#elif AIHC_GC == AIHC_GC_SEMISPACE
-  size_t bytes = sizeof(AihcSlot) * words;
-  if (machine->heap_next + bytes > machine->heap_limit) {
-    aihc_collect(machine, words, root_count, roots);
-  }
-#else
-#error "unknown AIHC_GC selection"
-#endif
+  aihc_gc_ensure(machine, words, root_count, roots);
 }
 
 AihcValue *aihc_make_node_unchecked(AihcMachine *machine, uint64_t tag,
                                     const AihcInfo *info) {
   uint64_t words = aihc_object_words(tag, info);
-  AihcValue *value;
-#if AIHC_GC == AIHC_GC_CALLOC
-  value = aihc_allocate_zeroed(sizeof(AihcSlot) * words);
-#elif AIHC_GC == AIHC_GC_SEMISPACE
-  size_t bytes = sizeof(AihcSlot) * words;
-  if (machine->heap_next + bytes > machine->heap_limit) {
-    aihc_fail("unchecked allocation exceeded reserved heap");
-  }
-  value = (AihcValue *)machine->heap_next;
-  machine->heap_next += bytes;
-  memset(value, 0, bytes);
-#else
-#error "unknown AIHC_GC selection"
-#endif
+  AihcValue *value = aihc_gc_allocate(machine, words);
   aihc_record_allocation(machine);
   value->header = aihc_make_header(tag, info);
   return value;
@@ -409,8 +198,8 @@ void aihc_reset_allocation_count(AihcMachine *machine) {
   machine->allocation_count = 0;
 }
 
-static const AihcInfo *aihc_next_application_info(const AihcInfo *info,
-                                                  uint64_t supplied_count) {
+const AihcInfo *aihc_next_application_info(const AihcInfo *info,
+                                           uint64_t supplied_count) {
   const AihcInfo *next = info->next;
   if (info->remaining_arity == 0 || next == NULL ||
       next->remaining_arity + 1 != info->remaining_arity ||
@@ -467,56 +256,6 @@ static AihcValue *aihc_copy_with_fields(AihcMachine *machine,
             : fields[index];
   }
   return copy;
-}
-
-static AihcSlot *aihc_portable_arguments(AihcMachine *machine,
-                                         AihcValue *function, uint64_t count,
-                                         const AihcSlot *values,
-                                         AihcValue *continuation) {
-  uint64_t field_count = aihc_value_info_table(function)->field_count;
-  size_t maximum_count = SIZE_MAX / sizeof(AihcSlot);
-  if (field_count > maximum_count || count > maximum_count - field_count) {
-    aihc_fail("runtime argument vector is too large");
-  }
-  size_t total = (size_t)field_count + (size_t)count;
-  if (continuation != NULL) {
-    if (total == maximum_count) {
-      aihc_fail("runtime argument vector is too large");
-    }
-    ++total;
-  }
-  AihcSlot *buffer =
-      aihc_reserve_slots(machine, &machine->portable_arguments,
-                         &machine->portable_arguments_capacity, total);
-  if (count != 0) {
-    memmove(buffer + field_count, values, sizeof(*values) * (size_t)count);
-  }
-  const AihcSlot *function_fields = aihc_value_fields_const(function);
-  if (field_count != 0) {
-    memcpy(buffer, function_fields,
-           sizeof(*function_fields) * (size_t)field_count);
-  }
-  if (continuation != NULL) {
-    buffer[(size_t)field_count + (size_t)count] = (AihcSlot)continuation;
-  }
-  return buffer;
-}
-
-static AihcPortableTransfer aihc_portable_transfer(AihcEntry entry,
-                                                   AihcSlot *arguments) {
-  return (AihcPortableTransfer){entry, arguments};
-}
-
-AihcPortableTransfer aihc_portable_call(AihcMachine *machine, AihcEntry entry,
-                                        uint64_t count,
-                                        const AihcSlot *arguments) {
-  AihcSlot *buffer =
-      aihc_reserve_slots(machine, &machine->portable_arguments,
-                         &machine->portable_arguments_capacity, count);
-  if (count != 0) {
-    memmove(buffer, arguments, sizeof(*arguments) * (size_t)count);
-  }
-  return aihc_portable_transfer(entry, buffer);
 }
 
 static AihcThread *aihc_thread_new(AihcMachine *machine) {
@@ -610,20 +349,9 @@ AihcMachine *aihc_machine_new(uint64_t global_count) {
   machine->globals = aihc_allocate_auxiliary(
       machine,
       sizeof(*machine->globals) * (global_count == 0 ? 1 : global_count));
-#if AIHC_GC == AIHC_GC_CALLOC
-  machine->heap_next = NULL;
-  machine->heap_limit = (uint8_t *)UINTPTR_MAX;
-#elif AIHC_GC == AIHC_GC_SEMISPACE
-  machine->semispace_bytes = AIHC_SEMISPACE_BYTES;
-  machine->heap_start =
-      aihc_allocate_auxiliary(machine, machine->semispace_bytes);
-  machine->other_space =
-      aihc_allocate_auxiliary(machine, machine->semispace_bytes);
-  machine->heap_next = machine->heap_start;
-  machine->heap_limit = machine->heap_start + machine->semispace_bytes;
-#endif
+  aihc_gc_init(machine);
   machine->current_thread = aihc_thread_new(machine);
-  machine->io_backend = aihc_default_io_backend();
+  machine->io_backend = aihc_host_io_backend();
   return machine;
 }
 
@@ -633,36 +361,6 @@ AihcSlot *aihc_alloc_locals(AihcMachine *machine, uint64_t count) {
 }
 
 void aihc_no_match(void) { aihc_fail("no matching case alternative"); }
-
-static AihcPortableTransfer
-aihc_portable_continue_values_now(AihcMachine *machine, AihcValue *continuation,
-                                  uint64_t count, const AihcSlot *values) {
-  if (continuation == NULL ||
-      aihc_value_tag(continuation) != AIHC_TAG_CLOSURE) {
-    aihc_fail("attempted to invoke a non-continuation value");
-  }
-  if (aihc_value_arity(continuation) != 1) {
-    aihc_fail("continuation closure does not accept exactly one result");
-  }
-  (void)aihc_next_application_info(aihc_value_info_table(continuation), count);
-  return aihc_portable_transfer(
-      aihc_value_entry(continuation),
-      aihc_portable_arguments(machine, continuation, count, values, NULL));
-}
-
-AihcPortableTransfer aihc_portable_continue_values(AihcMachine *machine,
-                                                   AihcValue *continuation,
-                                                   uint64_t count,
-                                                   const AihcSlot *values) {
-  return aihc_portable_continue_values_now(machine, continuation, count,
-                                           values);
-}
-
-static AihcPortableTransfer
-aihc_portable_continue_value(AihcMachine *machine, AihcValue *continuation,
-                             AihcSlot value) {
-  return aihc_portable_continue_values_now(machine, continuation, 1, &value);
-}
 
 AihcValue *aihc_apply_slow(AihcMachine *machine, AihcValue *function,
                            uint64_t count, const AihcSlot *arguments,
@@ -692,27 +390,6 @@ AihcValue *aihc_apply_slow(AihcMachine *machine, AihcValue *function,
   default:
     aihc_fail("attempted to apply a non-function value");
   }
-}
-
-AihcPortableTransfer aihc_portable_apply_cps(AihcMachine *machine,
-                                             AihcValue *function,
-                                             uint64_t count,
-                                             const AihcSlot *arguments,
-                                             AihcValue *continuation) {
-  if (function == NULL) {
-    aihc_fail("attempted to apply null");
-  }
-  if (aihc_value_tag(function) == AIHC_TAG_CLOSURE &&
-      aihc_value_arity(function) == 1) {
-    (void)aihc_next_application_info(aihc_value_info_table(function), count);
-    return aihc_portable_transfer(aihc_value_entry(function),
-                                  aihc_portable_arguments(machine, function,
-                                                          count, arguments,
-                                                          continuation));
-  }
-  AihcValue *applied =
-      aihc_apply_slow(machine, function, count, arguments, &continuation);
-  return aihc_portable_continue_value(machine, continuation, (AihcSlot)applied);
 }
 
 static void aihc_suspend_apply(AihcThread *thread, AihcValue *function,
@@ -756,12 +433,10 @@ static const AihcResume *aihc_select_thread(AihcMachine *machine,
   return resume;
 }
 
-static int32_t aihc_io_error(int error) {
-  return (int32_t)(-((int64_t)error) - 1);
-}
+int32_t aihc_io_error(int error) { return (int32_t)(-((int64_t)error) - 1); }
 
-static void aihc_resume_io_request(AihcMachine *machine, AihcIoRequest *request,
-                                   int32_t result) {
+void aihc_resume_io_request(AihcMachine *machine, AihcIoRequest *request,
+                            int32_t result) {
   AihcThread *thread = request->thread;
   AihcValue *continuation = request->continuation;
   request->state = AIHC_IO_COMPLETED;
@@ -773,179 +448,17 @@ static void aihc_resume_io_request(AihcMachine *machine, AihcIoRequest *request,
   aihc_enqueue_thread(machine, thread);
 }
 
-#ifndef AIHC_WASIP3
-static int aihc_posix_descriptor(const AihcIoHandle *handle) {
-  return (int)handle->backend_token;
-}
-
-static int aihc_posix_prepare(AihcIoRequest *request) {
-  int descriptor = aihc_posix_descriptor(request->handle);
-  int flags = fcntl(descriptor, F_GETFL);
-  if (flags == -1) {
-    return errno;
-  }
-  if ((flags & O_NONBLOCK) == 0 &&
-      fcntl(descriptor, F_SETFL, flags | O_NONBLOCK) == -1) {
-    return errno;
-  }
-  return 0;
-}
-
-static int aihc_posix_try_request(AihcIoRequest *request, int32_t *result) {
-  for (;;) {
-    ssize_t transferred;
-    uint8_t *bytes = request->buffer + request->offset;
-    if (request->kind == AIHC_IO_READ) {
-      transferred =
-          read(aihc_posix_descriptor(request->handle), bytes, request->length);
-    } else {
-      transferred =
-          write(aihc_posix_descriptor(request->handle), bytes, request->length);
-    }
-    if (transferred >= 0) {
-      *result = (int32_t)transferred;
-      return 1;
-    }
-    if (errno == EINTR) {
-      continue;
-    }
-    if (errno == EAGAIN || errno == EWOULDBLOCK) {
-      return 0;
-    }
-    *result = aihc_io_error(errno);
-    return 1;
-  }
-}
-
-static void aihc_complete_all_io_with_error(AihcMachine *machine, int error) {
-  AihcIoRequest *request = machine->io_requests_head;
-  machine->io_requests_head = NULL;
-  machine->io_requests_tail = NULL;
-  machine->io_request_count = 0;
-  while (request != NULL) {
-    AihcIoRequest *next = request->next;
-    aihc_resume_io_request(machine, request, aihc_io_error(error));
-    request = next;
-  }
-}
-
-static void aihc_posix_poll(AihcMachine *machine, int may_block) {
-  if (machine->io_request_count == 0) {
-    return;
-  }
-  size_t count = (size_t)machine->io_request_count;
-  struct pollfd *descriptors =
-      aihc_allocate_auxiliary(machine, sizeof(*descriptors) * count);
-  size_t index = 0;
-  for (AihcIoRequest *request = machine->io_requests_head; request != NULL;
-       request = request->next) {
-    descriptors[index].fd = aihc_posix_descriptor(request->handle);
-    descriptors[index].events =
-        request->kind == AIHC_IO_READ ? POLLIN : POLLOUT;
-    ++index;
-  }
-  int ready = poll(descriptors, count, may_block ? -1 : 0);
-  if (ready == -1) {
-    int error = errno;
-    free(descriptors);
-    if (error != EINTR) {
-      aihc_complete_all_io_with_error(machine, error);
-    }
-    return;
-  }
-  if (ready == 0) {
-    free(descriptors);
-    return;
-  }
-
-  AihcIoRequest **link = &machine->io_requests_head;
-  AihcIoRequest *tail = NULL;
-  index = 0;
-  while (*link != NULL) {
-    AihcIoRequest *request = *link;
-    short events = descriptors[index++].revents;
-    int32_t result = 0;
-    int complete = 0;
-    if ((events & POLLNVAL) != 0) {
-      result = aihc_io_error(EBADF);
-      complete = 1;
-    } else if (events != 0) {
-      complete = aihc_posix_try_request(request, &result);
-    }
-    if (complete) {
-      *link = request->next;
-      --machine->io_request_count;
-      aihc_resume_io_request(machine, request, result);
-    } else {
-      tail = request;
-      link = &request->next;
-    }
-  }
-  machine->io_requests_tail = tail;
-  free(descriptors);
-}
-
-static const AihcIoBackend aihc_posix_io_backend = {
-    aihc_posix_prepare,
-    aihc_posix_try_request,
-    aihc_posix_poll,
-};
-
-static const AihcIoBackend *aihc_default_io_backend(void) {
-  return &aihc_posix_io_backend;
-}
-#else
-extern int32_t aihc_wasip3_start_write(const unsigned char *bytes,
-                                       size_t length);
-
-static int aihc_wasip3_prepare(AihcIoRequest *request) {
-  (void)request;
-  return 0;
-}
-
-static int aihc_wasip3_try_request(AihcIoRequest *request, int32_t *result) {
-  if (request->kind != AIHC_IO_WRITE || request->handle->backend_token != 1) {
-    *result = aihc_io_error(1);
-    return 1;
-  }
-  int32_t status = aihc_wasip3_start_write(request->buffer + request->offset,
-                                           request->length);
-  if (status == INT32_MIN) {
-    return 0;
-  }
-  *result = status;
-  return 1;
-}
-
-static void aihc_wasip3_poll(AihcMachine *machine, int may_block) {
-  (void)machine;
-  (void)may_block;
-}
-
-static const AihcIoBackend aihc_wasip3_io_backend = {
-    aihc_wasip3_prepare,
-    aihc_wasip3_try_request,
-    aihc_wasip3_poll,
-};
-
-static const AihcIoBackend *aihc_default_io_backend(void) {
-  return &aihc_wasip3_io_backend;
-}
-#endif
-
 static const AihcResume *aihc_schedule(AihcMachine *machine) {
   for (;;) {
-    machine->io_backend->poll(machine, 0);
+    (void)machine->io_backend->poll(machine, 0);
     if (machine->run_queue_head != NULL) {
       return aihc_select_thread(machine, aihc_dequeue_thread(machine));
     }
     if (machine->io_request_count != 0) {
-      machine->io_backend->poll(machine, 1);
-#ifdef AIHC_WASIP3
-      return NULL;
-#else
+      if (machine->io_backend->poll(machine, 1) == AIHC_IO_POLL_SUSPENDED) {
+        return NULL;
+      }
       continue;
-#endif
     }
     aihc_fail("no runnable threads");
   }
@@ -974,10 +487,6 @@ static AihcIoRequest *aihc_io_submit(AihcIoKind kind, AihcIoHandle *handle,
   request->length = (size_t)length;
   return request;
 }
-
-void *aihc_io_stdin(void) { return &aihc_standard_input; }
-
-void *aihc_io_stdout(void) { return &aihc_standard_output; }
 
 static size_t aihc_byte_array_size(int64_t requested_size) {
   if (requested_size < 0 || (uint64_t)requested_size > SIZE_MAX) {
@@ -1297,35 +806,10 @@ const AihcResume *aihc_block_on_blackhole(AihcMachine *machine,
   return aihc_schedule(machine);
 }
 
-static AihcPortableTransfer aihc_portable_resume(AihcMachine *machine,
-                                                 const AihcResume *resume) {
-  if (resume == NULL) {
-    return (AihcPortableTransfer){0};
-  }
-  AihcPortableTransfer transfer;
-  switch (resume->kind) {
-  case AIHC_RESUME_APPLY:
-    transfer = aihc_portable_apply_cps(machine, resume->function, 0, NULL,
-                                       resume->continuation);
-    break;
-  case AIHC_RESUME_CONTINUE:
-    transfer = aihc_portable_continue_values_now(
-        machine, resume->function, resume->count,
-        resume->count == 0 ? NULL : &resume->value);
-    break;
-  default:
-    aihc_fail("invalid suspended continuation");
-  }
-  machine->selected_resume = (AihcResume){0};
-  return transfer;
-}
-
-#ifdef AIHC_WASIP3
-AihcPortableTransfer aihc_wasip3_complete_io(AihcMachine *machine,
-                                             int32_t result) {
+const AihcResume *aihc_complete_io(AihcMachine *machine, int32_t result) {
   AihcIoRequest *request = machine->io_requests_head;
   if (request == NULL) {
-    aihc_fail("WASI P3 completion has no pending request");
+    aihc_fail("IO completion has no pending request");
   }
   machine->io_requests_head = request->next;
   --machine->io_request_count;
@@ -1343,39 +827,7 @@ AihcPortableTransfer aihc_wasip3_complete_io(AihcMachine *machine,
   if (machine->io_requests_head == NULL) {
     machine->io_requests_tail = NULL;
   }
-  return aihc_portable_resume(machine, aihc_schedule(machine));
-}
-#endif
-
-AihcPortableTransfer aihc_portable_eval_cps(AihcMachine *machine,
-                                            AihcValue *value,
-                                            uint64_t result_is_lifted,
-                                            AihcValue *continuation,
-                                            AihcValue *update_continuation) {
-  if (value == NULL) {
-    aihc_fail("attempted to evaluate null");
-  }
-  switch (aihc_value_tag(value)) {
-  case AIHC_TAG_THUNK: {
-    AihcSlot *arguments =
-        aihc_portable_arguments(machine, value, 0, NULL, update_continuation);
-    AihcEntry entry = aihc_value_entry(value);
-    aihc_begin_blackhole(machine, value);
-    return aihc_portable_transfer(entry, arguments);
-  }
-  case AIHC_TAG_INDIRECTION:
-    if (result_is_lifted) {
-      return aihc_portable_eval_cps(machine, (AihcValue *)value->fields[0], 1,
-                                    continuation, update_continuation);
-    }
-    return aihc_portable_continue_value(machine, continuation,
-                                        value->fields[0]);
-  case AIHC_TAG_BLACKHOLE:
-    return aihc_portable_resume(
-        machine, aihc_block_on_blackhole(machine, value, continuation));
-  default:
-    return aihc_portable_continue_value(machine, continuation, (AihcSlot)value);
-  }
+  return aihc_schedule(machine);
 }
 
 void aihc_update(AihcValue *object, AihcValue *value) {
@@ -1440,64 +892,3 @@ void aihc_set_thread_done_continuation(AihcMachine *machine,
 }
 
 AihcEntry aihc_halt(AihcMachine *machine) { return machine->exit_code; }
-
-AihcPortableTransfer aihc_portable_fork_cps(AihcMachine *machine,
-                                            AihcValue *action,
-                                            AihcValue *continuation) {
-  AihcSlot thread_id = aihc_fork(machine, action);
-  return aihc_portable_continue_value(machine, continuation, thread_id);
-}
-
-AihcPortableTransfer aihc_portable_new_mvar_cps(AihcMachine *machine,
-                                                AihcValue *continuation) {
-  return aihc_portable_continue_value(machine, continuation,
-                                      (AihcSlot)aihc_mvar_new(machine));
-}
-
-AihcPortableTransfer aihc_portable_read_mvar_cps(AihcMachine *machine,
-                                                 void *mvar,
-                                                 AihcValue *continuation) {
-  return aihc_portable_resume(machine,
-                              aihc_mvar_read(machine, mvar, continuation));
-}
-
-AihcPortableTransfer aihc_portable_take_mvar_cps(AihcMachine *machine,
-                                                 void *mvar,
-                                                 AihcValue *continuation) {
-  return aihc_portable_resume(machine,
-                              aihc_mvar_take(machine, mvar, continuation));
-}
-
-AihcPortableTransfer aihc_portable_put_mvar_cps(AihcMachine *machine,
-                                                void *mvar, AihcValue *value,
-                                                AihcValue *continuation) {
-  return aihc_portable_resume(
-      machine, aihc_mvar_put(machine, mvar, (AihcSlot)value, continuation));
-}
-
-AihcPortableTransfer aihc_portable_yield_cps(AihcMachine *machine,
-                                             AihcValue *continuation) {
-  return aihc_portable_resume(machine, aihc_yield(machine, continuation));
-}
-
-AihcPortableTransfer aihc_portable_await_io_cps(AihcMachine *machine,
-                                                void *request,
-                                                AihcValue *continuation) {
-  return aihc_portable_resume(machine,
-                              aihc_await_io(machine, request, continuation));
-}
-
-AihcPortableTransfer aihc_portable_thread_done(AihcMachine *machine) {
-  return aihc_portable_resume(machine, aihc_thread_done(machine));
-}
-
-AihcPortableTransfer aihc_portable_start(AihcMachine *machine, AihcValue *root,
-                                         AihcValue *continuation,
-                                         AihcValue *update_continuation,
-                                         AihcValue *thread_done_continuation,
-                                         AihcEntry exit_code) {
-  machine->exit_code = exit_code;
-  aihc_set_thread_done_continuation(machine, thread_done_continuation);
-  return aihc_portable_eval_cps(machine, root, 1, continuation,
-                                update_continuation);
-}
