+++ components/aihc-native/test/Spec.hs
@@ -3,8 +3,8 @@ module Main (main) where
 import Test.Native.BlockLayout qualified as BlockLayout
 import Test.Native.Primitive qualified as Primitive
 import Test.Native.RegisterAllocate (tests)
+import Test.Native.RuntimePlan qualified as RuntimePlan
 import Test.Tasty (defaultMain, testGroup)
-import Test.Tasty.QuickCheck qualified as QC
 
 main :: IO ()
 main =
@@ -13,11 +13,7 @@ main =
         "aihc-native"
         [ BlockLayout.tests,
           Primitive.tests,
-          tests,
-          QC.testProperty "dummy quickcheck property" prop_dummy
+          RuntimePlan.tests,
+          tests
         ]
     )
-
--- | Keep the workspace-wide QuickCheck controls accepted by this suite.
-prop_dummy :: Bool -> Bool
-prop_dummy _ = True
