+++ components/aihc-wasm/src/Aihc/Wasm.hs
@@ -7,6 +7,7 @@ module Aihc.Wasm
     validatePrimitiveNames,
     validateProgramPrimitives,
     wasip3RuntimeSourcePath,
+    wasip3RuntimeSourcePaths,
     wasip3WorldPath,
   )
 where
@@ -16,6 +17,15 @@ import Paths_aihc_wasm (getDataFileName)
 
 wasip3RuntimeSourcePath :: IO FilePath
 wasip3RuntimeSourcePath = getDataFileName "runtime/aihc_wasip3.c"
+
+wasip3RuntimeSourcePaths :: IO [FilePath]
+wasip3RuntimeSourcePaths =
+  mapM
+    getDataFileName
+    [ "runtime/aihc_wasm_libc.c",
+      "runtime/aihc_wasm_adapter.c",
+      "runtime/aihc_wasip3.c"
+    ]
 
 wasip3WorldPath :: IO FilePath
 wasip3WorldPath = getDataFileName "runtime/wit"
