+++ scripts/nix/checks.nix
@@ -196,12 +196,12 @@
     mkdir -p "$bindings_directory"
     wit-bindgen c --world command --out-dir "$bindings_directory" components/aihc-wasm/runtime/wit
     while IFS= read -r -d "" file; do
-      if [[ "$file" == *components/aihc-wasm/runtime/aihc_wasip3.c ]]; then
+      if [[ "$file" == *components/aihc-wasm/runtime/*.c || "$file" == *aihc_host_wasip3.c ]]; then
         clang-tidy-unwrapped --quiet "$file" -- \
           --target=wasm32-unknown-unknown \
           -std=c11 -ffreestanding -Wall -Wextra -Wpedantic \
-          -DAIHC_WASIP3 \
           -Icomponents/aihc-wasm/runtime/include \
+          -Icomponents/aihc-wasm/runtime \
           -Icomponents/aihc-native/runtime \
           -isystem "$bindings_directory"
       else
